some types were int insteat of long (now less warnings on alpha).

+++ include/sys/types.h
@@ -52,7 +52,7 @@ typedef unsigned short mode_t;		/* Used for some file attributes. */
 typedef unsigned short nlink_t;		/* Used for link counts. */
 typedef signed long off_t;		/* Used for file sizes. */
 typedef signed int pid_t;		/* Used for process IDs and process group IDs. */
-typedef signed int ssize_t;		/* Used for a count of bytes or an error indication. */
+typedef signed long ssize_t;		/* Used for a count of bytes or an error indication. */
 typedef signed long suseconds_t;	/* Used for time in microseconds. */
 typedef signed long time_t;		/* Used for time in seconds. */
 typedef unsigned short uid_t;		/* Used for user IDs. */
